cloudrun/auto-cut/requirements.txt
moviepy==1.0.3
flask
flask-cors
google-cloud-storage
numpy